+++ internal/adapters/rest/auth/auth_test.go
@@ -27,7 +27,7 @@ func TestAuthenticate_Happy(t *testing.T) {
 	ctx := contextutil.WithReqID(context.Background(), 1)
 	auth := Authenticate(http.HandlerFunc(dummyHandler))
 	r := httptest.NewRequest("GET", "/", nil).WithContext(ctx)
-	claims := customClaims{UserID: 1, IsAdmin: false, RegisteredClaims: jwt.RegisteredClaims{ExpiresAt: jwt.NewNumericDate(time.Now().Add(5 * time.Minute))}}
+	claims := CustomClaims{UserID: 1, IsAdmin: false, RegisteredClaims: jwt.RegisteredClaims{ExpiresAt: jwt.NewNumericDate(time.Now().Add(5 * time.Minute))}}
 	token, err := jwt.NewWithClaims(jwt.SigningMethodHS512, claims).SignedString(jwtSecret)
 	assert.NoError(t, err)
 	r.Header.Set("Authorization", token)
@@ -42,7 +42,7 @@ func TestAuthenticate_NoExpiration(t *testing.T) {
 	ctx := contextutil.WithReqID(context.Background(), 1)
 	auth := Authenticate(http.HandlerFunc(dummyHandler))
 	r := httptest.NewRequest("GET", "/", nil).WithContext(ctx)
-	claims := customClaims{UserID: 1, IsAdmin: false}
+	claims := CustomClaims{UserID: 1, IsAdmin: false}
 	token, err := jwt.NewWithClaims(jwt.SigningMethodHS512, claims).SignedString(jwtSecret)
 	assert.NoError(t, err)
 	r.Header.Set("Authorization", token)
