+++ internal/adapters/rest/handlers/util.go
