+++ internal/adapters/rest/handlers/util_test.go
