+++ internal/adapters/rest/logger.go
@@ -1,4 +1,4 @@
-package web
+package rest
 
 import (
 	"log/slog"
