+++ internal/adapters/rest/logger_test.go
@@ -1,4 +1,4 @@
-package web
+package rest
 
 import (
 	"testing"
