+++ internal/adapters/rest/routes_test.go
@@ -1,4 +1,4 @@
-package web
+package rest
 
 import (
 	"context"
@@ -8,9 +8,8 @@ import (
 	"yadro-go-course/config"
 )
 
-func TestServer_SetupRoutes(t *testing.T) {
-	srv := NewServer()
-	srv.SetupRoutes(&config.Config{
+func TestApi(t *testing.T) {
+	Api(nil, nil, nil, &config.Config{
 		Server: config.Server{
 			RateLimit:        1,
 			ConcurrencyLimit: 1,
