+++ internal/adapters/rest/server.go
@@ -1,4 +1,4 @@
-package web
+package rest
 
 import (
 	"context"
